Move Month into a separate module
